refactor(config-ui): move @/components/utils/request to @/utils/request
