+++ config-ui/src/error/components/offline/api.ts
@@ -16,6 +16,6 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const ping = () => request('/ping')
