+++ config-ui/src/pages/blueprint/create/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const createBlueprint = (payload: any) =>
   request('/blueprints', {
