+++ config-ui/src/pages/blueprint/detail/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const getBlueprint = (id: ID) => request(`/blueprints/${id}`)
