+++ config-ui/src/pages/blueprint/home/api.ts
@@ -16,6 +16,6 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const getBlueprints = () => request('/blueprints')
