+++ config-ui/src/pages/pipeline/detail/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const getPipeline = (id: ID) => request(`/pipelines/${id}`)
