+++ config-ui/src/pages/project/detail/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const getProject = (name: string) => request(`/projects/${name}`)
