+++ config-ui/src/pages/transformation/detail/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 import { Plugins } from '@/plugins'
 
 export const createTransformation = (plugin: Plugins, payload: any) =>
