+++ config-ui/src/plugins/jira/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export type GetBoardsParams = {
   startAt: number
