+++ config-ui/src/plugins/webook/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 export const getConnections = () => request('/plugins/webhook/connections')
