+++ config-ui/src/store/connections/api.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import request from '@/components/utils/request'
+import { request } from '@/utils'
 
 type TestConnectionPayload = {
   endpoint: string
