+++ config-ui/src/utils/index.ts
@@ -16,5 +16,6 @@
  *
  */
 
+export * from './request'
 export * from './operator'
 export * from './time'
